% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marglik.R
\name{marglik}
\alias{marglik}
\title{Marginal likelihood of of the replication effect estimate}
\usage{
marglik(tr, sr, to, so, w = NULL, x = 1, y = 1, m = 0, v = 1)
}
\arguments{
\item{tr}{Effect estimate from the replication study.}

\item{sr}{Standard error of the replication effect estimate.}

\item{to}{Effect estimate from the original study.}

\item{so}{Standard error of the original effect estimate.}

\item{w}{Fixed weight parameter. Is only taken into account when not
\code{NULL}. Defaults to \code{NULL}.}

\item{x}{Number of successes parameter of the beta prior for \code{w}. Only
taken into account when \code{w} is \code{NULL}. Defaults to \code{1}.}

\item{y}{Number of failures parameter of the beta prior for \code{w}. Only
taken into account when \code{w} is \code{NULL}. Defaults to \code{1}.}

\item{m}{Mean parameter of the normal prior component. Defaults to \code{0}.}

\item{v}{Variance parameter of the normal prior component. Defaults to
\code{1}.}
}
\value{
The marginal likelihood of the replication effect estimate.
}
\description{
This function computes the marginal likelihood of the
    replication data under a mixture prior that incorporates the original
    data. See the details section for details regarding data model and prior
    distributions.
}
\details{
A normal likelihood around the underlying effect size \eqn{\theta}
    is assumed for the effect estimate from the replication study
\deqn{\code{tr} \mid \theta \sim \text{N}(\theta, \code{sr}^2)}{tr | theta ~
    N(theta, sr^2)}

    A mixture prior is assumed for the effect size \eqn{\theta} with one
    component being a normal distribution with mean equal to the original
    effect estimate \code{to}, variance equal to the squared original
    standard error \code{so^2}, and mixture weight \code{w}, and the other
    component being a normal distribution with mean \code{m}, variance
    \code{v}, and mixture weight \eqn{1 - \code{w}}
\deqn{\theta \mid w \sim \code{w} \times \text{N}(\code{to}, \code{so}^2) + (1 - \code{w})
    \times \text{N}(\code{m}, \code{v})}{ theta | \code{w} ~ \code{w} * N(to, so^2) + (1 -
    \code{w}) * N(m, v)}

    The mixture weight \code{w} can either be fixed to a value between zero
and one, or a Beta prior can be assumed for it
\deqn{w \sim \text{Beta}(x, y)}{w ~ Beta(x, y)}
}
\examples{
## uniform prior specified for weight w
marglik(tr = 0.09, sr = 0.05, to = 0.21, so = 0.05, x = 1, y = 1, m = 0, v = 4)
marglik(tr = 0.21, sr = 0.06, to = 0.21, so = 0.05, x = 1, y = 1, m = 0, v = 4)
marglik(tr = 0.44, sr = 0.04, to = 0.21, so = 0.05, x = 1, y = 1, m = 0, v = 4)

## fixed weight w = 0.5
marglik(tr = 0.09, sr = 0.05, to = 0.21, so = 0.05, w = 0.2, m = 0, v = 4)
marglik(tr = 0.21, sr = 0.06, to = 0.21, so = 0.05, w = 0.2, m = 0, v = 4)
marglik(tr = 0.44, sr = 0.04, to = 0.21, so = 0.05, w = 0.2, m = 0, v = 4)

}
\author{
Samuel Pawel, Roberto Macri Demmartino
}
