% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{normal2mixposterior}
\alias{normal2mixposterior}
\title{Posterior based on two component normal mixture prior}
\usage{
normal2mixposterior(y, se, m1, v1, m2, v2, w)
}
\arguments{
\item{y}{Data mean.}

\item{se}{Standard error of the data mean.}

\item{m1}{Mean parameter of the first normal prior component.}

\item{v1}{Variance parameter of first the normal prior component.}

\item{m2}{Mean parameter of the second normal prior component.}

\item{v2}{Variance parameter of second the normal prior component.}

\item{w}{Mixture weight.}
}
\value{
A list of the posterior means, variances, and weight.
}
\description{
This function computes the posterior means, variances, and the
    weight parameter based on data with normal likelihood and a two component
    mixture prior for the underlying mean.
}
\examples{
## normal2mixposterior(y = 0.09, se = 0.05, m1 = 0.21, v1 = 0.05^2,
##                     m2 = 0, v2 = 1, w = 0.5)

}
\author{
Samuel Pawel
}
\keyword{internal}
