% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posteriormix}
\alias{posteriormix}
\title{Posterior density of effect size and weight parameter}
\usage{
posteriormix(theta, w, tr, sr, to, so, x = 1, y = 1, m = 0, v = 1)
}
\arguments{
\item{theta}{Effect size. Has to be of length one or the same length as
\code{w}.}

\item{w}{Weight parameter. Has to be of length one or the same length as
\code{theta}.}

\item{tr}{Effect estimate from the replication study.}

\item{sr}{Standard error of the replication effect estimate.}

\item{to}{Effect estimate from the original study.}

\item{so}{Standard error of the original effect estimate.}

\item{x}{Number of successes parameter of the beta prior for \code{w}.
Defaults to \code{1}.}

\item{y}{Number of failures parameter of the beta prior for \code{w}.
Defaults to \code{1}.}

\item{m}{Mean parameter of the normal prior component. Defaults to \code{0}.}

\item{v}{Variance parameter of the normal prior component. Defaults to
\code{1}.}
}
\value{
The posterior density for a pair of effect size and weight parameter.
}
\description{
This function computes the posterior density of effect size and
    weight parameter based on the data from original and replication study.
    See the details section for details regarding data model and prior
    distributions.
}
\details{
A normal likelihood around the underlying effect size \eqn{\theta}
    is assumed for the effect estimate from the replication study
    \deqn{\code{tr} \mid \theta \sim \text{N}(\theta, \code{sr}^2)}{tr |
    theta ~ N(theta, sr^2)}

    A mixture prior is specified for the effect size \eqn{\theta} with one
    component being a normal distribution with mean equal to the original
    effect estimate \code{to}, variance equal to the squared original
    standard error \code{so^2}, and mixture weight \code{w}, and the other
    component being a normal distribution with mean \code{m}, variance
    \code{v}, and mixture weight 1 - \code{w}
\deqn{\theta \mid w \sim w
    \times \text{N}(\code{to}, \code{so}^2) + (1 - w) \times
    \text{N}(\code{m}, \code{v})}{ theta | w ~ w * N(to, so^2) + (1 - w) *
    N(m, v)}

    A Beta prior is specified for the weight parameter
\deqn{w \sim \text{Beta}(x, y)}{w ~ Beta(x, y)}
}
\examples{
## 2D plot of posterior density
w <- seq(0, 1, length.out = 200)
theta <- seq(0, 0.3, length.out = 200)
parGrid <- expand.grid(w = w, theta = theta)
postdens <- posteriormix(theta = parGrid$theta, w = parGrid$w, tr = 0.09,
                    sr = 0.05, to = 0.2, so = 0.05, m = 0, v = 2)
postdensMat <- matrix(data = postdens, ncol = 200, byrow = TRUE)
filled.contour(x = theta, y = w, z = postdensMat,
               xlab = bquote("Effect size" ~ theta),
               ylab = "Weight parameter w", nlevels = 15,
               color.palette = function(n) hcl.colors(n = n, palette = "Blues 3", rev = TRUE))

}
\author{
Samuel Pawel, Roberto Macri Demmartino
}
