% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{thetaHPD}
\alias{thetaHPD}
\title{Highest posterior density interval for effect size}
\usage{
thetaHPD(level = 0.95, tr, sr, to, so, w = NULL, x = 1, y = 1, m = 0, v = 1)
}
\arguments{
\item{level}{Credibility level. Defaults to \code{0.95}.}

\item{tr}{Effect estimate from the replication study.}

\item{sr}{Standard error of the replication effect estimate.}

\item{to}{Effect estimate from the original study.}

\item{so}{Standard error of the original effect estimate.}

\item{w}{Fixed weight parameter. Is only taken into account when not
\code{NULL}. Defaults to \code{NULL}.}

\item{x}{Number of successes parameter of the beta prior for \code{w}. Only
taken into account when \code{w} is \code{NULL}. Defaults to \code{1}.}

\item{y}{Number of failures parameter of the beta prior for \code{w}. Only
taken into account when \code{w} is \code{NULL}. Defaults to \code{1}.}

\item{m}{Mean parameter of the normal prior component. Defaults to \code{0}.}

\item{v}{Variance parameter of the normal prior component. Defaults to
\code{1}.}
}
\value{
The highest posterior density interval at the specified credibility
    level and posterior median for the effect size.
}
\description{
This function computes the highest posterior density interval
    for the effect size based on the data from original and replication study
    using a mixture prior to incorporate the original data into the analysis
    of the replication data. See the details section for details regarding
    data model and prior distributions.
}
\details{
A normal likelihood around the underlying effect size \eqn{\theta}
    is assumed for the effect estimate from the replication study
\deqn{\code{tr} \mid \theta \sim \text{N}(\theta, \code{sr}^2)}{tr | theta ~
    N(theta, sr^2)}

    A mixture prior is assumed for the effect size \eqn{\theta} with one
    component being a normal distribution with mean equal to the original
    effect estimate \code{to}, variance equal to the squared original
    standard error \code{so^2}, and mixture weight \eqn{w}, and the other
    component being a normal distribution with mean \code{m}, variance
    \code{v}, and mixture weight \eqn{1 - w}
\deqn{\theta \mid w \sim w \times \text{N}(\code{to}, \code{so}^2) + (1 - w)
    \times \text{N}(\code{m}, \code{v})}{ theta | w ~ w * N(to, so^2) + (1 -
    w) * N(m, v)}

    The weight parameter can either be fixed to a specified value or a Beta
prior can assumed for it \deqn{w \sim \text{Beta}(x, y)}{w ~ Beta(x, y)}
}
\examples{
## tipping point analysis (computing HPDs for range of weights from 0 to 1)
to <- 0.21
tr <- 0.09
so <- 0.05
sr <- 0.045
w <- seq(0, 1, length.out = 20)
HPDs <- t(sapply(X = w, FUN = function(w) {
               thetaHPD(level = 0.95, tr = tr, sr = sr, to = to,
                        so = so, w = w, m = 0, v = 4)
}))
plot(NA, NA, type = "n", ylim = c(0, 0.3), xlim = c(-0.25, 1.25),
     xlab = "Weight of original study",
     ylab = "Posterior median with 95\% HPDI", las = 1)
arrows(x0 = w, y0 = HPDs[,1], y1 = HPDs[,3], code = 3, angle = 90, length = 0.05)
points(x = w, y = HPDs[,2], pch = 20)
arrows(x0 = c(-0.2, 1.2), y0 = c(tr - 1.96*sr, to - 1.96*so),
       y1 = c(tr + 1.96*sr, to + 1.96*so), code = 3, angle = 90, length = 0.05)
points(x = c(-0.2, 1.2), y = c(tr, to), pch = 20)
axis(side = 1, at = c(-0.2, 1.2), labels = c("Replication", "Original"))

}
\author{
Samuel Pawel, Roberto Macri Demmartino
}
