% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{wHPD}
\alias{wHPD}
\title{Highest posterior density interval for effect size}
\usage{
wHPD(level = 0.95, tr, sr, to, so, x = 1, y = 1, m = 0, v = 1)
}
\arguments{
\item{level}{Credibility level. Defaults to \code{0.95}.}

\item{tr}{Effect estimate from the replication study.}

\item{sr}{Standard error of the replication effect estimate.}

\item{to}{Effect estimate from the original study.}

\item{so}{Standard error of the original effect estimate.}

\item{x}{Number of successes parameter of the beta prior for \code{w}.
Defaults to \code{1}.}

\item{y}{Number of failures parameter of the beta prior for \code{w}.
Defaults to \code{1}.}

\item{m}{Mean parameter of the normal prior component. Defaults to \code{0}.}

\item{v}{Variance parameter of the normal prior component. Defaults to
\code{1}.}
}
\value{
The highest posterior density interval at the specified credibility
    level and posterior median for the mixture weight.
}
\description{
This function computes the highest posterior density interval
    for the effect size based on the data from original and replication study
    using a mixture prior to incorporate the original data into the analysis
    of the replication data. See the details section for details regarding
    data model and prior distributions.
}
\details{
A normal likelihood around the underlying effect size \eqn{\theta}
    is assumed for the effect estimate from the replication study
\deqn{\code{tr} \mid \theta \sim \text{N}(\theta, \code{sr}^2)}{tr | theta ~
    N(theta, sr^2)}

    A mixture prior is assumed for the effect size \eqn{\theta} with one
    component being a normal distribution with mean equal to the original
    effect estimate \code{to}, variance equal to the squared original
    standard error \code{so^2}, and mixture weight \eqn{w}, and the other
    component being a normal distribution with mean \code{m}, variance
    \code{v}, and mixture weight \eqn{1 - w}
\deqn{\theta \mid w \sim w \times \text{N}(\code{to}, \code{so}^2) + (1 - w)
    \times \text{N}(\code{m}, \code{v})}{ theta | w ~ w * N(to, so^2) + (1 -
    w) * N(m, v)}

    A Beta prior is assumed for the weight parameter \deqn{w \sim
\text{Beta}(x, y)}{w ~ Beta(x, y)}
}
\examples{
wHPD(level = 0.95, tr = 0.09, sr = 0.05, to = 0.21, so = 0.05, x = 1, y = 1,
      m = 0, v = 4)
wHPD(level = 0.95, tr = 0.21, sr = 0.06, to = 0.21, so = 0.05, x = 1, y = 1,
      m = 0, v = 4)
wHPD(level = 0.95, tr = 0.44, sr = 0.04, to = 0.21, so = 0.05, x = 1, y = 1,
      m = 0, v = 4)

}
\author{
Samuel Pawel, Roberto Macri Demmartino
}
